requirements.txt
matplotlib==3.2.1
numpy==1.19.5
Pillow==8.0.1
tensorflow==2.4.0
